+++ NEWS.md
@@ -1,3 +1,17 @@
+# rstatix 0.6.0.999
+
+## New features
+
+
+## Major changes
+
+
+## Minor changes
+
+
+## Bug fixes
+   
+
 # rstatix 0.6.0
    
 ## Minor changes